+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 * Added `ExtractCELLxGENEMeta` to extract metadata of CELLxGENE datasets.
 * Added `ParseCELLxGENE` to download objects from CELLxGENE.
+* Added `ExtractHCAMeta` to extract metadata of Human Cell Atlas projects.
 
 -------------------